+++ chemclipse/plugins/org.eclipse.chemclipse.chromatogram.msd.identifier/src/org/eclipse/chemclipse/chromatogram/msd/identifier/settings/IDatabaseIdentifierSettings.java
@@ -22,7 +22,7 @@ public interface IDatabaseIdentifierSettings extends IIdentifierSettings {
 
 	void setRetentionTimeWindowForDatabase(int retentionTimeWindow);
 
-	float getRetentionIndexWindowForDatabase();
+	int getRetentionIndexWindowForDatabase();
 
-	void setRetentionIndexWindowForDatabase(float retentionIndexWindow);
+	void setRetentionIndexWindowForDatabase(int retentionIndexWindow);
 }
